rename redis.ts to cache.ts
