+++ src/commands-message/general/config.ts
@@ -1,5 +1,5 @@
 import type { MessageCommand } from "@type/commands";
-import { slashCommandIdsCache } from "@utils/redis";
+import { slashCommandIdsCache } from "@utils/cache";
 
 export default {
     name: "config",
