+++ src/index.ts
@@ -1,7 +1,7 @@
 import { initializeDatabase, client } from "./utils/initalize";
 import { logger } from "./utils/logger";
 import { getAccessToken } from "./utils/osu";
-import { initializeRedis, closeRedis } from "./utils/redis";
+import { initializeRedis, closeRedis } from "./utils/cache";
 import { createHandler } from "@lilybird/handlers/simple";
 import { CachingDelegationType, createClient, Intents } from "lilybird";
 import { Channel, Guild, GuildVoiceChannel } from "@lilybird/transformers";
