+++ src/listeners/guildCreate.ts
@@ -3,7 +3,7 @@ import { logger } from "@utils/logger";
 import { Tables } from "@type/database";
 import type { Guild } from "@type/database";
 import type { Event } from "@lilybird/handlers";
-import { guildPrefixesCache } from "@utils/redis";
+import { guildPrefixesCache } from "@utils/cache";
 
 export default {
     event: "guildCreate",
